🎨 feat(logo): unify logo across Toolbar + Preloader, drop "WEB" sub
Problem: The Toolbar used "Sonic Pi" + a small uppercase "WEB" subscript
in dimmed comment color, while the Preloader splash used a different
title (large pink "Sonic Pi Web" without the icon). Two surfaces, two
visual treatments — the user noticed and asked for a single
consistent wordmark.

Fix: extract createLogo() into a shared `src/app/Logo.ts` helper that
returns the canonical mark — pink ♫ glyph with theme.accentHover glow
+ "Sonic Pi Web" in theme.fg, bold, 0.95 rem, 0.5 px letter-spacing.
Both Toolbar and Preloader now consume this helper, so any future
visual tweak lands in both places automatically.

Drops the small "WEB" subscript from the toolbar — it was redundant
once the logo text became "Sonic Pi Web" in full.

The preloader bumps the logo slightly (icon 1.8 rem, text 1.15 rem)
since it's the splash centerpiece, but the structure is identical.
